+++ tee-worker/client-api-api/src/identity/interfaces/vc/definitions.ts
