+++ tee-worker/client-api-api/src/omni/interfaces/sidechain/definitions.ts
