+++ tee-worker/client-api-api/src/omni/interfaces/trusted_operations/definitions.ts
