+++ tee-worker/client-api-api/src/omni/interfaces/vc/definitions.ts
